+++ CHANGES.md
@@ -1,6 +1,11 @@
 # CHANGES
 ## RosarioSIS Student Information System
 
+Changes in 7.2.4
+----------------
+- Take in Account Calendar Day Minutes in UpdateAttendanceDaily.fnc.php
+- Return false if School Periods Length sum is 0 in UpdateAttendanceDaily.fnc.php, thanks to @dzungdo
+
 Changes in 7.2.3
 ----------------
 - Fix regression since 5.9 search text User Field in Search.fnc.php, thanks to @dzungdo